Do not emit a spurious warning when "directory" argument to "rmdir -p" is absolute, with one or more leading slash.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@137902 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ bin/rmdir/rmdir.c
@@ -103,9 +103,11 @@ rm_path(char *path)
 	*++p = '\0';
 	while ((p = strrchr(path, '/')) != NULL) {
 		/* Delete trailing slashes. */
-		while (--p > path && *p == '/')
+		while (--p >= path && *p == '/')
 			;
 		*++p = '\0';
+		if (p == path)
+			break;
 
 		if (rmdir(path) < 0) {
 			warn("%s", path);
